matrix algorithm done
